requirements.txt
Flask==2.3.2
pandas==2.0.3
psycopg2-binary==2.9.6
scikit-learn==1.3.0
joblib==1.3.1
